Revamp editor toolbar UI and percent formatting
Introduce a redesigned EditorToolbar: add a centered AI action button with breathing glow and sparkles, reorganize formatting controls into left/center/right groups, implement animated transitions and haptic feedback, and simplify pop-out menus with rounded shapes (removed harsh top border). Also adjust toolbar sizing/keyboard behavior and refactor toggle/button interactions. Minor fix in HomeScreen to format progress as an integer percentage. Add implementation plan, task list, and walkthrough artifact docs.

+++ app/src/main/java/com/denser/june/presentation/screens/home/HomeScreen.kt
@@ -209,7 +209,7 @@ fun DayProgressIndicator() {
     val minutes = (displaySeconds % 3600) / 60
     val seconds = displaySeconds % 60
     val formattedTime = String.format(Locale.getDefault(), "%02d:%02d:%02d", hours, minutes, seconds)
-    val formattedPercent = String.format(Locale.getDefault(), "%.2f%%", displayProgress * 100)
+    val formattedPercent = String.format(Locale.getDefault(), "%d%%", (displayProgress * 100).toInt())
 
     Box(
         modifier = Modifier
